Use the filter based syncing flow
Discussed on dev calls, it seems the approach for recovery is to let the
user decide how many scripts they have used for an old wallet. Without
using lossy methods, this is about as best we can do. This is now
reflected in `ScanType::Recovery`

On the API re-design, we now check the filters as they are emitted by
the node, queue them up, and fetch them at the end.