Update `bip157 = 0.2.0`
The `Node` is now as I/O minimal as possible, at least when it comes to
disk. The current local chain is capable of reconsiling reorgs, but we
will have to use every emitted `Header`.

This reduces the disk footprint from 80 bytes * `block height` to 32
bytes * `block height`, and the `Wallet` acts as the source of truth for
the headers. In general this speeds up syncs a lot.